move shared functions to utils.py
