+++ src/class_final_eval.py
@@ -6,15 +6,13 @@ Authors: Kenneth Schackart
 
 import argparse
 import os
-from typing import Any, BinaryIO, List, NamedTuple, TextIO, Tuple, cast
+from typing import BinaryIO, List, NamedTuple, TextIO
 
-import torch
-from datasets import load_metric
 from torch.utils.data.dataloader import DataLoader
-from transformers import AutoModelForSequenceClassification as classifier
 
 from class_data_handler import DataFields, RunParams, get_dataloader
-from utils import CustomHelpFormatter, Metrics, get_torch_device
+from utils import (CustomHelpFormatter, get_classif_metrics, get_classif_model,
+                   get_torch_device, save_metrics)
 
 
 # ---------------------------------------------------------------------------
@@ -106,30 +104,6 @@ def get_args():
 
 
 # ---------------------------------------------------------------------------
-def get_model(checkpoint_fh: BinaryIO,
-              device: torch.device) -> Tuple[Any, str]:
-    """
-    Instatiate predictive model from checkpoint
-
-    Params:
-    `checkpoint_fh`: Model checkpoint filehandle
-    `device`: The `torch.device` to use
-
-    Return:
-    Model instance from checkpoint, and model name
-    """
-
-    checkpoint = torch.load(checkpoint_fh, map_location=device)
-    model_name = checkpoint['model_name']
-    model = classifier.from_pretrained(model_name, num_labels=2)
-    model.load_state_dict(checkpoint['model_state_dict'])
-    model.to(device)
-    model.eval()
-
-    return model, model_name
-
-
-# ---------------------------------------------------------------------------
 def get_test_dataloader(args: Args, model_name: str) -> DataLoader:
     """
     Generate the dataloaders
@@ -156,65 +130,6 @@ def get_test_dataloader(args: Args, model_name: str) -> DataLoader:
 
 
 # ---------------------------------------------------------------------------
-def get_metrics(model: Any, dataloader: DataLoader,
-                device: torch.device) -> Metrics:
-    """
-    Compute model performance metrics
-
-    Parameters:
-    `model`: Classification model
-    `dataloader`: DataLoader containing tokenized text entries and
-    corresponding labels
-    `device`: Torch device
-
-    Return:
-    A `Metrics` NamedTuple
-    """
-    calc_precision = load_metric('precision')
-    calc_recall = load_metric('recall')
-    calc_f1 = load_metric('f1')
-    total_loss = 0.
-    num_seen_datapoints = 0
-    for batch in dataloader:
-        batch = {k: v.to(device) for k, v in batch.items()}
-        with torch.no_grad():
-            outputs = model(**batch)
-        num_seen_datapoints += len(batch['input_ids'])
-        predictions = torch.argmax(outputs.logits, dim=-1)
-        calc_precision.add_batch(predictions=predictions,
-                                 references=batch['labels'])
-        calc_recall.add_batch(predictions=predictions,
-                              references=batch['labels'])
-        calc_f1.add_batch(predictions=predictions, references=batch['labels'])
-        total_loss += outputs.loss.item()
-    total_loss /= num_seen_datapoints
-
-    precision = cast(dict, calc_precision.compute())
-    recall = cast(dict, calc_recall.compute())
-    f1 = cast(dict, calc_f1.compute())
-
-    return Metrics(precision['precision'], recall['recall'], f1['f1'],
-                   total_loss)
-
-
-# ---------------------------------------------------------------------------
-def save_metrics(metrics: Metrics, filename: str) -> None:
-    """
-    Save test metrics to text file
-
-    Parameters:
-    `metrics`: A `Metrics` NamedTuple
-    """
-
-    with open(filename, 'wt') as fh:
-        print('precision,recall,f1,loss', file=fh)
-        print(f'{metrics.precision},{metrics.recall},',
-              f'{metrics.f1},{metrics.loss}',
-              sep='',
-              file=fh)
-
-
-# ---------------------------------------------------------------------------
 def main() -> None:
     """ Main function """
 
@@ -228,11 +143,11 @@ def main() -> None:
 
     device = get_torch_device()
 
-    model, model_name = get_model(args.checkpoint, device)
+    model, model_name = get_classif_model(args.checkpoint, device)
 
     dataloader = get_test_dataloader(args, model_name)
 
-    test_metrics = get_metrics(model, dataloader, device)
+    test_metrics = get_classif_metrics(model, dataloader, device)
 
     save_metrics(test_metrics, out_file)
